+++ files/initrd/opt/rr/include/modules.sh
@@ -5,6 +5,11 @@
 function getAllModules() {
   PLATFORM=${1}
   KVER=${2}
+
+  if [ -z "${PLATFORM}" -o -z "${KVER}" ]; then
+    echo ""
+    return 1
+  fi
   # Unzip modules for temporary folder
   rm -rf "${TMP_PATH}/modules"
   mkdir -p "${TMP_PATH}/modules"
@@ -29,6 +34,11 @@ function addToModules() {
   PLATFORM=${1}
   KVER=${2}
   KOFILE=${3}
+
+  if [ -z "${PLATFORM}" -o -z "${KVER}" -o -z "${KOFILE}" ]; then
+    echo ""
+    return 1
+  fi
   # Unzip modules for temporary folder
   rm -rf "${TMP_PATH}/modules"
   mkdir -p "${TMP_PATH}/modules"
@@ -46,6 +56,11 @@ function delToModules() {
   PLATFORM=${1}
   KVER=${2}
   KONAME=${3}
+
+  if [ -z "${PLATFORM}" -o -z "${KVER}" -o -z "${KONAME}" ]; then
+    echo ""
+    return 1
+  fi
   # Unzip modules for temporary folder
   rm -rf "${TMP_PATH}/modules"
   mkdir -p "${TMP_PATH}/modules"
@@ -74,6 +89,11 @@ function getdepends() {
   PLATFORM=${1}
   KVER=${2}
   KONAME=${3}
+
+  if [ -z "${PLATFORM}" -o -z "${KVER}" -o -z "${KONAME}" ]; then
+    echo ""
+    return 1
+  fi
   # Unzip modules for temporary folder
   rm -rf "${TMP_PATH}/modules"
   mkdir -p "${TMP_PATH}/modules"
